I screwed up with the directories when I moved this from private to public.
This commit fixes the problem.
